Enemy/bossManta: body drawMantaShadow (0.2% -> 99.8%)
Reconstruct the two-pass shadow silhouette render: draw each active
enemy's shape packets twice (front-cull additive, back-cull
subtractive blend) into the alpha channel, EFB-copy that mask to an
A8/I8 texture, blend the manta/angry-manta color (TBossMantaParams
LERP by unk84/15) into a KColor, then draw a full-screen quad with
GXSetTexCoordGen2/alpha-compare TEV ops to composite the shadow.

Retyped TBossMantaManager::unk80 to JUtility::TColor (was s32; byte
writes at +0/1/2/3 confirm a packed GXColor) and
TMantaBattleState::unk4 to s32 (was u8; word-load in target asm).

Residual is pure stack-slot renumbering (frame size differs by a
constant, same MWCC do-last artifact as the rest of the TU).

+++ include/Enemy/BossManta.hpp
@@ -4,6 +4,7 @@
 #include <Enemy/Enemy.hpp>
 #include <Enemy/EnemyManager.hpp>
 #include <Strategic/Nerve.hpp>
+#include <JSystem/JUtility/JUTColor.hpp>
 
 class TBossManta : public TSpineEnemy {
 public:
@@ -113,7 +114,7 @@ public:
 	public:
 		void update();
 		TBossMantaManager* unk0;
-		u8 unk4;
+		s32 unk4;
 	};
 
 	TBossMantaManager(const char*);
@@ -135,7 +136,7 @@ public:
 	/* 0x74 */ JGeometry::TVec3<f32>* unk74;
 	/* 0x78 */ JGeometry::TVec3<f32>* unk78;
 	/* 0x7C */ u8* unk7C;
-	/* 0x80 */ s32 unk80;
+	/* 0x80 */ JUtility::TColor unk80;
 	/* 0x84 */ s32 unk84;
 	/* 0x88 */ TMantaBattleState unk88;
 	/* 0x90 */ TMantaMessageState unk90;
